--- signup_signin/signup_signin/user/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-
-from .models import UserProfile
-
-admin.site.register(UserProfile)
