--- signup_signin/signup_signin/user/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class UserConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'user'
