--- signup_signin/signup_signin/user/templates/greet.html
@@ -1,27 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}
-    Greeting
-{% endblock %}
-
-{% block body %}
-<div class="container mt-5">
-    <div class="row justify-content-center">
-        <div class="col-md-6">
-            <div class="card">
-                <div class="card-header">
-                    Welcome, {{ request.user.full_name }}!
-                </div>
-                <div class="card-body">
-                    <p class="card-text">Thank you for signing up. We're excited to have you onboard!</p>
-                    {% comment %} <a href="{% url 'home' %}" class="btn btn-primary">Start Exploring</a> {% endcomment %}
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-<div class="text-center mt-3">
-    <a href="{% url 'logout' %}" class="btn btn-secondary">Logout</a>
-    <a href="{% url 'delete' %}" class="btn btn-secondary">Delete account</a>
-</div>
-{% endblock %}
