--- signup_signin/signup_signin/user/templates/login.html
@@ -1,34 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}
-    {% if title %}
-        {{ title }}
-    {% endif %}
-{% endblock %}
-
-{% block body %}
-
-<div class="d-flex justify-content-center" style="min-height: 100vh;">
-    <div class="col-md-6">
-        {% if title %}
-            <h2 class="text-center">{{ title }}</h2>
-        {% endif %}
-        <form method="POST" action="."> {% csrf_token %}
-            <div class="mb-3">
-                <label for="{{ form.username.id_for_label }}" class="form-label">Username</label>
-                {{ form.username }}
-            </div>
-            <div class="mb-3">
-                <label for="{{ form.password.id_for_label }}" class="form-label">Password</label>
-                {{ form.password }}
-            </div>
-            <div class="text-center">
-                <button type="submit" class="btn btn-primary">Submit</button>
-            </div>
-        </form>
-        {% if errors %}
-            {{ errors }}
-        {% endif %}
-    </div>
-</div>
-{% endblock %}
